Add a CheckedScenario type which always validates
